move test utils into "common" folder [URO-190]
